+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## UNRELEASED
+*April 13, 2023*
+
+## v1.0.7
 
 ### Improvements